requirements.txt
mysql-connector==2.2.9
aniso8601==8.0.0
asgiref==3.2.5
astroid==2.3.3
click==7.1.1
colorama==0.4.3
Django==3.0.4
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
SQLAlchemy==1.3.15
sqlparse==0.3.1
mysql-connector-python==8.0.5
werkzeug==0.16.1 
Flask-Cors==3.0.8        
Flask-Login==0.5.0        
flask-oidc==1.4.0        
Flask-RESTful==0.3.8        
flask-restplus==0.9.2 
